+++ SME_Implementations/SingleCycleRISCV/Incrementer_Buses.cs
@@ -5,7 +5,7 @@ namespace SingleCycleRISCV {
     [TopLevelOutputBus]
     public interface Incrementer_Output : IBus {
         [InitialValue]
-        uint Address {get; set;}
+        ulong Address {get; set;}
     }
 
 }
